requirements.txt
requests>=2.31.0
pdfplumber>=0.10.0
streamlit>=1.30.0
supabase>=2.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
pandas>=2.0.0
